Refactor query engine configuration and remove BasicNarrowSolver
* Updated default solver in configuration to KeyValueStoreSolver across various files.
* Removed BasicNarrowSolver implementation and its associated tests.
* Adjusted documentation to reflect changes in solver defaults and requirements for project_id and container_tags_table.
* Ensured compatibility of report generation with the new solver setup.
* Updated integration tests to utilize KeyValueStoreSolver instead of the removed BasicNarrowSolver.
