+++ tests/data/config/config.json
@@ -18,7 +18,7 @@
     ]
   },
   "query_engine": {
-    "solver": "BasicNarrowSolver"
+    "solver": "KeyValueStoreSolver"
   },
   "measurement_dimensions": [
     "uut_id",
