+++ tests/mda_query_engine/unit/analyze/full_test.py
@@ -1,7 +1,7 @@
 from pyspark.sql.types import BinaryType
 
-from mda_query_engine.analyze.query.solvers.basic_narrow_solver import BasicNarrowSolver
 from mda_query_engine.analyze.query.solvers.delta_solver import DeltaSolver
+from mda_query_engine.analyze.query.solvers.key_value_store_solver import KeyValueStoreSolver
 from mda_query_engine.model.series.sample_series import SampleSeries
 
 
@@ -27,7 +27,7 @@ def test_basic_narrow_full(spark, basic_narrow_db):
     expr1 = c1.max().alias("eng_rpm_max")
     expr2 = c2.max().alias("veh_spd_max")
 
-    df = query.select(expr1, expr2).solve(spark, solver=BasicNarrowSolver(spark))
+    df = query.select(expr1, expr2).solve(spark, solver=KeyValueStoreSolver(spark))
     assert df.count() == 3
     assert df.select("eng_rpm_max").collect()[0]["eng_rpm_max"] >= 0
 
@@ -38,7 +38,7 @@ def test_sample_series_returned_as_sample_series(spark, basic_narrow_db):
     rpm = query.channel(channel_name="Engine RPM")
     expr = rpm.alias("rpm_series")
 
-    pdf = query.select(expr).toPandas(spark, solver=BasicNarrowSolver(spark))
+    pdf = query.select(expr).toPandas(spark, solver=KeyValueStoreSolver(spark))
     assert len(pdf) == 3
 
     for _, row in pdf.iterrows():
@@ -53,7 +53,7 @@ def test_sample_series_solve_returns_binary(spark, basic_narrow_db):
     rpm = query.channel(channel_name="Engine RPM")
     expr = rpm.alias("rpm_series")
 
-    df = query.select(expr).solve(spark, solver=BasicNarrowSolver(spark))
+    df = query.select(expr).solve(spark, solver=KeyValueStoreSolver(spark))
     assert df.count() == 3
 
     schema_field = df.schema["rpm_series"]
@@ -67,7 +67,7 @@ def test_sample_series_mixed_with_scalar(spark, basic_narrow_db):
     expr_series = rpm.alias("rpm_series")
     expr_max = rpm.max().alias("rpm_max")
 
-    pdf = query.select(expr_series, expr_max).toPandas(spark, solver=BasicNarrowSolver(spark))
+    pdf = query.select(expr_series, expr_max).toPandas(spark, solver=KeyValueStoreSolver(spark))
     assert len(pdf) == 3
 
     for _, row in pdf.iterrows():
