+++ tests/mda_reporting/integration/container_and_sequence_event_test.py
@@ -58,7 +58,7 @@ def test_container_event_in_report(spark, basic_narrow_db):
                 ]
             ]
         ),
-        query_engine=QueryEngine(solver=Solvers.BASIC_NARROW_SOLVER),
+        query_engine=QueryEngine(solver=Solvers.KEY_VALUE_STORE_SOLVER),
         measurement_dimensions=[
             MeasurementDimensions.CONTAINER_ID,
             MeasurementDimensions.START_TS,
@@ -169,7 +169,7 @@ def test_container_event_with_basic_event(spark, basic_narrow_db):
                 ]
             ]
         ),
-        query_engine=QueryEngine(solver=Solvers.BASIC_NARROW_SOLVER),
+        query_engine=QueryEngine(solver=Solvers.KEY_VALUE_STORE_SOLVER),
         measurement_dimensions=[
             MeasurementDimensions.CONTAINER_ID,
             MeasurementDimensions.START_TS,
@@ -254,7 +254,7 @@ def test_sequence_of_events_without_max_overlap_in_report(spark, basic_narrow_db
                 ]
             ]
         ),
-        query_engine=QueryEngine(solver=Solvers.BASIC_NARROW_SOLVER),
+        query_engine=QueryEngine(solver=Solvers.KEY_VALUE_STORE_SOLVER),
         measurement_dimensions=[
             MeasurementDimensions.CONTAINER_ID,
             MeasurementDimensions.START_TS,
