+++ tests/mda_reporting/unit/aggregations/histogram_test.py
@@ -1,7 +1,7 @@
 import pytest
 
 from mda_query_engine.analyze.metadata.time_series_expression import TimeSeriesSelector
-from mda_query_engine.analyze.query.solvers.basic_narrow_solver import BasicNarrowSolver
+from mda_query_engine.analyze.query.solvers.key_value_store_solver import KeyValueStoreSolver
 from mda_reporting.aggregations.histogram import (
     HistogramCustomWeights,
     HistogramDistance,
@@ -179,7 +179,7 @@ def test_determine_aggregations(spark, basic_narrow_db):
     eng_rpm = basic_narrow_db.query.channel(channel_name="Engine RPM")
     event = BasicEvent(name="test_event_1", expr=eng_rpm > 1000)
     hist = HistogramDuration(name="test_hist", base_expr=eng_rpm, event=event, bins=[0.0, 10000.0])
-    solver = BasicNarrowSolver(spark)
+    solver = KeyValueStoreSolver(spark)
     solved_df = basic_narrow_db.query.select(hist.get_expression()).solve(spark, solver)
     df = HistogramDuration.determine_aggregations(
         spark=spark, aggregations=[hist], solved_df=solved_df
@@ -527,7 +527,7 @@ def test_histogram_custom_weights_determine_aggregations(spark, basic_narrow_db)
         event=event,
         bins=[0.0, 10000.0],
     )
-    solver = BasicNarrowSolver(spark)
+    solver = KeyValueStoreSolver(spark)
     solved_df = basic_narrow_db.query.select(hist.get_expression()).solve(spark, solver)
     df = HistogramCustomWeights.determine_aggregations(
         spark=spark, aggregations=[hist], solved_df=solved_df
@@ -912,7 +912,7 @@ def test_histogram_distance_determine_aggregations(spark, basic_narrow_db):
         event=event,
         bins=[0.0, 10000.0],
     )
-    solver = BasicNarrowSolver(spark)
+    solver = KeyValueStoreSolver(spark)
     solved_df = basic_narrow_db.query.select(hist.get_expression()).solve(spark, solver)
     df = HistogramDistance.determine_aggregations(
         spark=spark, aggregations=[hist], solved_df=solved_df
